requirements.txt
selenium~=4.26.1
unittest2~=1.1.0
chromedriver-autoinstaller~=0.3.1
allure-python-commons~=2.9.45
pytest~=8.3.3